change chainId from Int/UInt to BigInt type Ethereum plugin

+++ packages/cli/src/__tests__/plugin/expected-types/types.ts
@@ -56,7 +56,7 @@ export interface Ethereum_TxRequest {
   gasPrice?: BigInt | null;
   data?: String | null;
   value?: BigInt | null;
-  chainId?: UInt32 | null;
+  chainId?: BigInt | null;
   type?: UInt32 | null;
 }
 
@@ -104,7 +104,7 @@ export interface Ethereum_EventNotification {
 /* URI: "ens/ethereum.web3api.eth" */
 export interface Ethereum_Network {
   name: String;
-  chainId: Int;
+  chainId: BigInt;
   ensAddress?: String | null;
 }
 
@@ -118,7 +118,7 @@ export interface Ethereum_TxResponse {
   gasPrice?: BigInt | null;
   data: String;
   value: BigInt;
-  chainId: UInt32;
+  chainId: BigInt;
   blockNumber?: BigInt | null;
   blockHash?: String | null;
   timestamp?: UInt32 | null;
